aficare-agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aficare-agent"
version = "0.1.0"
description = "AI-powered medical assistant for African healthcare settings"
authors = [{name = "AfiCare Team", email = "team@aficare.org"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "streamlit>=1.28.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0.1",
    "python-multipart>=0.0.6",
    "llama-cpp-python>=0.2.20",
    "sentence-transformers>=2.2.2",
    "scikit-learn>=1.3.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "cryptography>=41.0.0",
    "bcrypt>=4.1.0",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]

[project.urls]
Homepage = "https://github.com/aficare/aficare-agent"
Repository = "https://github.com/aficare/aficare-agent"
Documentation = "https://docs.aficare.org"

[project.scripts]
aficare = "src.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true